framework for redirecting to internal or external URLs depending on configartion array which is setup in conf.php
initial use will be for redirecting bookmarked URLs which are incompatible with the clean URL scheme. i still need some regex experts to suggest some standard data upon which to redirect

if you have access to apache rewrite, you should use that for internal URLs instead of this module. this module causes browsers to redirect whereas url rewrite does not.